Refactored commons names
