+++ dcs-use-case-commons/src/main/scala/it/unibo/dcs/commons/use_case/executor/executor.scala
@@ -4,7 +4,6 @@ import rx.lang.scala.Scheduler
 import rx.lang.scala.schedulers.ExecutionContextScheduler
 
 import scala.concurrent.ExecutionContext
-import scala.language.implicitConversions
 
 package object executor {
